\name{nwcorner-package}
\alias{nwcorner-package}
\alias{nwcorner}
\docType{package}
\title{
\packageTitle{nwcorner}
}
\description{
\packageDescription{nwcorner}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{nwcorner}
\packageIndices{nwcorner}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{nwcorner}

Maintainer: \packageMaintainer{nwcorner}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
